rename: InputType enum 폴더 위치 수정
